Improved: Parametric filter calculations (command "Filter") are now fully using double precision. This yields a great improvement in signal-to-noise ratio for practically no performance penalty. Fixed: The method BiQuad::gainAt was returning wrong values. As this method was only used for testing purposes, this change is not user-visible.
git-svn-id: svn://svn.code.sf.net/p/equalizerapo/code/trunk@42 177c5db2-39f6-4273-851a-b0754c4d5937

+++ filters/IIRFilter.h
@@ -33,11 +33,11 @@ public:
 
 private:
 	unsigned order;
-	float b0;
-	float* a;
-	float* b;
+	double b0;
+	double* a;
+	double* b;
 	unsigned channelCount;
-	float* x;
-	float* y;
+	double* x;
+	double* y;
 };
 #pragma AVRT_VTABLES_END
